+++ app/domain/auth/auth_service.py
@@ -1,21 +1,14 @@
 from sqlalchemy.orm import Session
-from datetime import datetime, timedelta, timezone
-from dotenv import load_dotenv
-import jwt
 import os
 
 from ...core.logger_config import setup_logger
-from ...lib.google_api import get_google_login_url, get_google_token, get_google_user_data
-from ...models.user import User
+from ...lib.google_api import get_google_login_url, get_google_token, get_google_user_data, decode_id_token
+from ...lib.jwt_util import make_access_token, decode_token
 from . import auth_crud
 from ..user import user_service
 
 
 logger = setup_logger()
-
-load_dotenv()
-JWT_SECRET = os.getenv('JWT_SECRET')
-JWT_ALGORITHM = os.getenv('JWT_ALGORITHM')
 
 
 def login_google() -> str:
@@ -46,15 +39,24 @@ def auth_google_access_token(access_token: str, db: Session):
 
     return make_access_token(db_user)
 
-def make_access_token(user: User):
-    return jwt.encode({
-        "sub": user.user_id,
-        "name": user.user_name,
-        "email": user.user_email,
-        "exp": datetime.now(timezone.utc) + timedelta(hours=1)
-    }, JWT_SECRET, algorithm=JWT_ALGORITHM)
+def auth_google_id_token(id_token: str, db: Session):
+    user_info = decode_id_token(id_token)
+
+    logger.info(f"📌 get user info to google id token - {user_info}")
+
+    # existing_user = user_service.get_user_by_email(user_data.email, db)
+    
+
+    # if existing_user:
+    #     db_user = existing_user
+    # else:
+    #     db_user = auth_crud.create_user(db, user_data)
+
+    #     logger.info(f"📌 successfully make user - {db_user}")
+
+    # logger.info(f"📌 login complete!")
+
+    # return make_access_token(db_user)
 
 def decode_token(token: str):
-    payload = jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
-    logger.info(f"📌 token - {payload}")
-    return payload
+    return decode_token(token)
