+++ app/schemas/auth/auth_request_schema.py
@@ -6,3 +6,6 @@ class LoginGoogleCode(BaseModel):
 
 class LoginGoogleToken(BaseModel):
     token: str
+
+class LoginGoogleIdToken(BaseModel):
+    id_token: str
